+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Changed OrderId to Reference for payment to be consistent with Fraud
+
 ## [1.1.14] - 2022-03-23
 
 ### Changed